Parameterize progress callgraph tests
