+++ tests/test_client.py
@@ -1,3 +1,4 @@
+from collections import defaultdict
 import git
 import os
 import pathlib
@@ -5,11 +6,13 @@ import sys
 import tempfile
 import toml
 
+import pytest
 import unittest
 
 from declib.artifacts import (
-    FunctionHeader, StackVariable, Comment, Struct
+    Function, FunctionHeader, StackVariable, Comment, Struct
 )
+from binsync.controller import BSController
 from binsync.core.client import Client
 
 
@@ -134,6 +137,95 @@ class TestClient(unittest.TestCase):
             # exception because the TOML fails to load
             self.assertRaises(toml.decoder.TomlDecodeError, lambda: Client("user0", tmpdir, "fake_hash"))
             
+
+class FakeDecompilerInterface:
+    name = "fake"
+    binary_hash = "fake_hash"
+    default_func_prefix = "sub_"
+
+    def __init__(self):
+        self.artifact_change_callbacks = defaultdict(list)
+        self.functions = {
+            0x400100: Function(addr=0x400100, size=0x20, name="renamed_func")
+        }
+        self.comments = {}
+        self.global_vars = {}
+        self.enums = {}
+        self.typedefs = {}
+        self.structs = {}
+        self.patches = {}
+        self.segments = {}
+
+    def get_func_size(self, addr):
+        func = self.functions.get(addr)
+        return func.size if func is not None else 0
+
+    def info(self, *_args, **_kwargs):
+        pass
+
+    def warning(self, *_args, **_kwargs):
+        pass
+
+    def shutdown(self):
+        pass
+
+    def start_artifact_watchers(self):
+        pass
+
+    def stop_artifact_watchers(self):
+        pass
+
+
+class FailingCallgraphDecompilerInterface(FakeDecompilerInterface):
+    def get_callgraph(self):
+        raise AttributeError("'int' object has no attribute 'function'")
+
+
+class CodeXrefDecompilerInterface(FakeDecompilerInterface):
+    def __init__(self):
+        super().__init__()
+        self.functions = {
+            0x400100: Function(addr=0x400100, size=0x20, name="caller"),
+            0x400200: Function(addr=0x400200, size=0x20, name="callee"),
+        }
+
+    def xrefs_to(self, artifact, decompile=False, only_code=False):
+        if artifact.addr == 0x400200 and only_code:
+            return [self.functions[0x400100]]
+        return []
+
+
+class TestProgressCallgraph:
+    @pytest.mark.parametrize(
+        ("decompiler_interface", "expected_nodes", "expected_edges"),
+        [
+            pytest.param(
+                FailingCallgraphDecompilerInterface,
+                {0x400100},
+                set(),
+                id="no-xrefs",
+            ),
+            pytest.param(
+                CodeXrefDecompilerInterface,
+                {0x400100, 0x400200},
+                {(0x400100, 0x400200)},
+                id="caller-to-callee-xref",
+            ),
+        ],
+    )
+    def test_progress_callgraph_fallback(
+        self, decompiler_interface, expected_nodes, expected_edges
+    ):
+        controller = BSController(
+            decompiler_interface=decompiler_interface(),
+            headless=True,
+        )
+
+        graph = controller.get_progress_callgraph()
+
+        assert {func.addr for func in graph.nodes} == expected_nodes
+        assert {(caller.addr, callee.addr) for caller, callee in graph.edges} == expected_edges
+
 
 if __name__ == "__main__":
     unittest.main(argv=sys.argv)
